Link to our version of the xmonad cheat sheet
Instead of the wiki version (which is semi-hard to change), we should
better link to a version of the image that we control completely.

For example, in 82b14896264b2ac22ec644355d5e31e61a431cb7 we switched the
IRC server in the image over to Libera, but the image on the wiki is
still not updated.

Related: https://github.com/xmonad/xmonad-web/pull/26