Add some keybinds and tidy up
Add keybind for showing default keybinds.
Do some minor cleanup of lines and things,
as long as I was in there already.
Add missing Xinerama keybinds for moving windows.
Add XMonad logo, as it was in the original png,
as it appeared on the haskell wiki.
Update http to https.

Fixes xmonad/xmonad#287