Overhaul quick start
The quick start, as it stood, did not make much sense.  The linked
tutorial already assumes that the user has installed xmonad-contrib as a
library, so this should be mentioned further up.

Further, at this point it is perhaps not advisable to link to the config
archive, which holds a lot of old configurations but basically no
up-to-date ones.  The tutorial already links to some maintainers'
configurations, which should be enough to get a glimpse of what's
possible.

In order to not get into trouble of alt vs. mod, the keybindings where
also changed to the more agnostic EZConfig-like syntax.